move logic from table component to useStore hook
